Graffiti not being actively developed. (#221)
With Graffiti not being maintained anymore, and GraphQL rapidly evolving and catching up to it's spec....it may not be an "awesome" tool anymore.  The example repo could still useful in some cases so......that being said, if the purpose of this list is just a catalog of absolutely everything GraphQL then disregard.